requirements.txt
PyQt5>=5.15.0
opencv-python>=4.9.0
winocr>=0.0.15
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.26.0
Pillow>=10.0.0
